// src/app/comics/interfaces/ComicRes.ts
// Generated by https://quicktype.io

import { Comic } from './Comic';

export interface ComicRes {
    code:            number;
    status:          string;
    copyright:       string;
    attributionText: string;
    attributionHTML: string;
    etag:            string;
    data:            Data;
}

export interface Data {
    offset:  number;
    limit:   number;
    total:   number;
    count:   number;
    results: Comic[];
}